+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryChildView.kt
@@ -1,0 +1,7 @@
+package org.simple.clinic.summary
+
+typealias PatientSummaryModelUpdateCallback = (PatientSummaryChildModel) -> Unit
+
+interface PatientSummaryChildView {
+  fun registerSummaryModelUpdateCallback(callback: PatientSummaryModelUpdateCallback?)
+}