+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryScreen.kt
@@ -73,7 +73,7 @@ import javax.inject.Named
 class PatientSummaryScreen(
     context: Context,
     attrs: AttributeSet
-) : RelativeLayout(context, attrs), PatientSummaryScreenUi, PatientSummaryUiActions {
+) : RelativeLayout(context, attrs), PatientSummaryScreenUi, PatientSummaryUiActions, PatientSummaryChildView {
 
   @Inject
   lateinit var screenRouter: ScreenRouter
@@ -96,6 +96,8 @@ class PatientSummaryScreen(
 
   @Inject
   lateinit var whatsAppMessageSender: WhatsAppMessageSender
+
+  private var modelUpdateCallback: PatientSummaryModelUpdateCallback? = null
 
   private val snackbarActionClicks = PublishSubject.create<PatientSummaryEvent>()
   private val events: Observable<PatientSummaryEvent> by unsafeLazy {
@@ -130,7 +132,10 @@ class PatientSummaryScreen(
         init = PatientSummaryInit(),
         update = PatientSummaryUpdate(),
         effectHandler = effectHandlerFactory.create(this).build(),
-        modelUpdateListener = viewRenderer::render
+        modelUpdateListener = { model ->
+          modelUpdateCallback?.invoke(model)
+          viewRenderer.render(model)
+        }
     )
   }
   private val teleconsultationErrorSnackbar by unsafeLazy {
@@ -165,6 +170,42 @@ class PatientSummaryScreen(
 
     val screenDestroys: Observable<ScreenDestroyed> = detaches().map { ScreenDestroyed() }
     alertFacilityChangeSheetClosed(screenDestroys)
+    setupChildViewVisibility(screenDestroys)
+  }
+
+  @SuppressLint("CheckResult")
+  private fun setupChildViewVisibility(
+      screenDestroys: Observable<ScreenDestroyed>
+  ) {
+    val modelUpdates: List<Observable<PatientSummaryChildModel>> =
+        listOf(
+            this,
+            drugSummaryView,
+            bloodPressureSummaryView,
+            bloodSugarSummaryView,
+            assignedFacilityView,
+            medicalHistorySummaryView
+        ).map(::createSummaryChildModelStream)
+
+    Observable
+        .combineLatest(modelUpdates) { models -> models.map { it as PatientSummaryChildModel } }
+        .filter { models -> models.all(PatientSummaryChildModel::readyToRender) }
+        .take(1)
+        .takeUntil(screenDestroys)
+        .subscribe {
+          summaryLoadingProgressBar.visibility = GONE
+          summaryViewsContainer.visibility = VISIBLE
+        }
+  }
+
+  private fun createSummaryChildModelStream(
+      summaryChildView: PatientSummaryChildView
+  ): Observable<PatientSummaryChildModel> {
+    return Observable.create { emitter ->
+      summaryChildView.registerSummaryModelUpdateCallback(emitter::onNext)
+
+      emitter.setCancellable { summaryChildView.registerSummaryModelUpdateCallback(null) }
+    }
   }
 
   override fun onAttachedToWindow() {
@@ -501,6 +542,10 @@ class PatientSummaryScreen(
 
   override fun hideAssignedFacilityView() {
     assignedFacilityView.visibility = View.GONE
+  }
+
+  override fun registerSummaryModelUpdateCallback(callback: PatientSummaryModelUpdateCallback?) {
+    modelUpdateCallback = callback
   }
 }