+++ app/src/main/java/org/simple/clinic/summary/assignedfacility/AssignedFacilityModel.kt
@@ -3,13 +3,14 @@ package org.simple.clinic.summary.assignedfacility
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 import org.simple.clinic.facility.Facility
+import org.simple.clinic.summary.PatientSummaryChildModel
 import java.util.UUID
 
 @Parcelize
 data class AssignedFacilityModel(
     val patientUuid: UUID,
     val assignedFacility: Facility?
-) : Parcelable {
+) : Parcelable, PatientSummaryChildModel {
 
   companion object {
     fun create(patientUuid: UUID) = AssignedFacilityModel(
@@ -20,6 +21,10 @@ data class AssignedFacilityModel(
 
   val hasAssignedFacility: Boolean
     get() = assignedFacility != null
+
+  override fun readyToRender(): Boolean {
+    return hasAssignedFacility
+  }
 
   fun assignedFacilityUpdated(facility: Facility?): AssignedFacilityModel {
     return copy(assignedFacility = facility)