+++ app/src/main/java/org/simple/clinic/summary/bloodsugar/BloodSugarSummaryViewModel.kt
@@ -3,7 +3,7 @@ package org.simple.clinic.summary.bloodsugar
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 import org.simple.clinic.bloodsugar.BloodSugarMeasurement
-import org.simple.clinic.facility.Facility
+import org.simple.clinic.summary.PatientSummaryChildModel
 import java.util.UUID
 
 @Parcelize
@@ -11,12 +11,16 @@ data class BloodSugarSummaryViewModel(
     val patientUuid: UUID,
     val measurements: List<BloodSugarMeasurement>?,
     val totalRecordedBloodSugarCount: Int?
-) : Parcelable {
+) : Parcelable, PatientSummaryChildModel {
 
   companion object {
     fun create(patientUuid: UUID): BloodSugarSummaryViewModel {
       return BloodSugarSummaryViewModel(patientUuid, null, null)
     }
+  }
+
+  override fun readyToRender(): Boolean {
+    return totalRecordedBloodSugarCount != null && measurements != null
   }
 
   fun summaryFetched(measurements: List<BloodSugarMeasurement>): BloodSugarSummaryViewModel {